Full hibernate(almost)

+++ dao/src/main/java/com/gmail/kurmazpavel/beans/ShippingItem.java
@@ -4,9 +4,9 @@ public class ShippingItem {
     private String name;
     private int amount;
     private double price;
-    private int order_id;
+    private long order_id;
 
-    public ShippingItem(String name, int amount, double price, int order_id) {
+    public ShippingItem(String name, int amount, double price, long order_id) {
         this.name = name;
         this.amount = amount;
         this.price = price;
@@ -25,7 +25,7 @@ public class ShippingItem {
         return price;
     }
 
-    public int getOrder_id(){return order_id;}
+    public long getOrder_id(){return order_id;}
 
     public void setName(String name) {
         this.name = name;
@@ -38,7 +38,8 @@ public class ShippingItem {
     public void setPrice(double price) {
         this.price = price;
     }
-    public void setOrder_id(int order_id) {this.order_id = order_id;}
+
+    public void setOrder_id(long order_id) {this.order_id = order_id;}
 
     @Override
     public String toString() {
